+++ bundle/src/main/java/com/adobe/acs/commons/httpcache/engine/impl/TeeServletOutputStream.java
@@ -105,7 +105,6 @@ public class TeeServletOutputStream extends ServletOutputStream {
             writeListener.onWritePossible();
         } catch (IOException e) {
             listener.onError(e);
-            e.printStackTrace();
         }
     }
 }