+++ bundle/src/main/java/com/adobe/acs/commons/httpcache/store/jcr/impl/JCRHttpCacheStoreImpl.java
@@ -21,38 +21,6 @@ package com.adobe.acs.commons.httpcache.store.jcr.impl;
 
 import com.adobe.acs.commons.functions.CheckedConsumer;
 import com.adobe.acs.commons.functions.CheckedFunction;
-import java.io.IOException;
-import java.util.Collections;
-import java.util.Dictionary;
-import java.util.Map;
-import java.util.concurrent.CopyOnWriteArrayList;
-
-import javax.jcr.Node;
-import javax.jcr.Session;
-import javax.management.NotCompliantMBeanException;
-import javax.management.openmbean.CompositeType;
-import javax.management.openmbean.OpenDataException;
-import javax.management.openmbean.OpenType;
-import javax.management.openmbean.SimpleType;
-
-import com.adobe.acs.commons.util.impl.exception.CacheMBeanException;
-import org.apache.commons.io.FileUtils;
-import org.apache.commons.io.IOUtils;
-
-import org.apache.sling.api.resource.ResourceResolver;
-import org.apache.sling.api.resource.ResourceResolverFactory;
-import org.apache.sling.commons.classloader.DynamicClassLoaderManager;
-import org.apache.sling.commons.osgi.PropertiesUtil;
-import org.osgi.service.component.ComponentContext;
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-import org.osgi.service.metatype.annotations.AttributeDefinition;
-import org.osgi.service.metatype.annotations.Designate;
-import org.osgi.service.metatype.annotations.ObjectClassDefinition;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import com.adobe.acs.commons.httpcache.config.HttpCacheConfig;
 import com.adobe.acs.commons.httpcache.engine.CacheContent;
 import com.adobe.acs.commons.httpcache.exception.HttpCacheDataStreamException;
@@ -73,29 +41,45 @@ import com.adobe.acs.commons.httpcache.store.jcr.impl.writer.EntryNodeWriter;
 import com.adobe.acs.commons.httpcache.store.mem.impl.MemTempSinkImpl;
 import com.adobe.acs.commons.util.impl.AbstractJCRCacheMBean;
 import com.adobe.acs.commons.util.impl.JcrCacheMBean;
+import com.adobe.acs.commons.util.impl.exception.CacheMBeanException;
+import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.IOUtils;
+import org.apache.sling.api.resource.ResourceResolver;
+import org.apache.sling.api.resource.ResourceResolverFactory;
+import org.apache.sling.commons.classloader.DynamicClassLoaderManager;
+import org.apache.sling.commons.osgi.PropertiesUtil;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+import org.osgi.service.metatype.annotations.Designate;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.jcr.Node;
+import javax.jcr.Session;
+import javax.management.NotCompliantMBeanException;
+import javax.management.openmbean.CompositeType;
+import javax.management.openmbean.OpenDataException;
+import javax.management.openmbean.OpenType;
+import javax.management.openmbean.SimpleType;
+import java.io.IOException;
+import java.util.Collections;
+import java.util.Map;
+import java.util.concurrent.CopyOnWriteArrayList;
 
 /**
  * ACS AEM Commons - HTTP Cache - JCR based cache store implementation.
  */
 
-@Component(service= {HttpCacheStore.class, JcrCacheMBean.class, Runnable.class}, property= {
-      HttpCacheStore.KEY_CACHE_STORE_TYPE + "=" +  HttpCacheStore.VALUE_JCR_CACHE_STORE_TYPE,
-      "jmx.objectname"  + "=" +  "com.adobe.acs.httpcache:type=JCR HTTP Cache Store",
-      "scheduler.concurrent" + "=" + "false",
-      JCRHttpCacheStoreImpl.PN_BUCKETDEPTH  + ":Integer=" + JCRHttpCacheStoreImpl.DEFAULT_BUCKETDEPTH
+@Component(service = {HttpCacheStore.class, JcrCacheMBean.class, Runnable.class}, property = {
+        HttpCacheStore.KEY_CACHE_STORE_TYPE + "=" + HttpCacheStore.VALUE_JCR_CACHE_STORE_TYPE,
+        "jmx.objectname" + "=" + "com.adobe.acs.httpcache:type=JCR HTTP Cache Store",
+        "scheduler.concurrent" + "=" + "false"
 
 })
-@Designate(ocd=JCRHttpCacheStoreImpl.Config.class)
+@Designate(ocd = Config.class)
 public class JCRHttpCacheStoreImpl extends AbstractJCRCacheMBean<CacheKey, CacheContent> implements HttpCacheStore, JcrCacheMBean, Runnable {
 
-    //property keys
-    public static final String PN_ROOTPATH = "httpcache.config.jcr.rootpath";
-    public static final String PN_BUCKETDEPTH = "httpcache.config.jcr.bucketdepth";
-    public static final String PN_SAVEDELTA = "httpcache.config.jcr.savedelta";
-    public static final String PN_EXPIRETIMEINSECONDS = "httpcache.config.jcr.expiretimeinseconds";
-
-    //defaults
-    public static final String DEFAULT_ROOTPATH = "/var/acs-commons/httpcache";
     private static final String SERVICE_NAME = "httpcache-jcr-storage-service";
 
     private static final Map<String, Object> AUTH_INFO;
@@ -104,59 +88,18 @@ public class JCRHttpCacheStoreImpl extends AbstractJCRCacheMBean<CacheKey, Cache
         AUTH_INFO = Collections.singletonMap(ResourceResolverFactory.SUBSERVICE, (Object) SERVICE_NAME);
     }
 
-    //By default, we go for the maximum bucket depth. This uses the full hashcode of 10 digits.
-    public static final int DEFAULT_BUCKETDEPTH = 10;
-
-    //Perform a save on a delta of 500 by default.
-    public static final int DEFAULT_SAVEDELTA = 500;
-
-    // 1 week.
-    public static final int DEFAULT_EXPIRETIMEINSECONDS = 604800;
-
     private static final Logger log = LoggerFactory.getLogger(JCRHttpCacheStoreImpl.class);
 
     //fields
     private String cacheRootPath;
     private int bucketTreeDepth;
     private int deltaSaveThreshold;
-    private int expireTimeInSeconds;
+    private long expireTimeInMilliSeconds;
 
 
-    @ObjectClassDefinition(name = "ACS AEM Commons - Http Cache - JCR Cache Store.",
-            description = "Cache data store implementation for JCR storage.")
-    public @interface Config {
-        @AttributeDefinition(
-              name = "Cache clean-up schedule",
-                description = "[every minute = 0 * * * * ?] Visit www.cronmaker.com to generate cron expressions.",
-                defaultValue = "0 0 12 1/1 * ? *"
-            )
-        String scheduler_expression();
-
-        @AttributeDefinition(
-                  name = "Cache-root Parent Path location",
-                description = "Points to the location of the cache root parent node in the JCR repository",
-                   defaultValue = JCRHttpCacheStoreImpl.DEFAULT_ROOTPATH
-            )
-        String httpcache_config_jcr_rootpath();
-        
-        @AttributeDefinition(
-                  name = "Save threshold",
-                description = "The threshold to add,remove and modify nodes when handling the cache",
-                defaultValue = ""+JCRHttpCacheStoreImpl.DEFAULT_SAVEDELTA
-            )
-        int httpcache_config_jcr_savedelta();
-        
-        @AttributeDefinition(
-                  name = "Expire time in seconds",
-                description = "The time seconds after which nodes will be removed by the scheduled cleanup service. ",
-                defaultValue = ""+JCRHttpCacheStoreImpl.DEFAULT_EXPIRETIMEINSECONDS
-            )
-        int httpcache_config_jcr_expiretimeinseconds();
-    }
-    
-    @Reference 
-    private ResourceResolverFactory   resourceResolverFactory;
-    @Reference 
+    @Reference
+    private ResourceResolverFactory resourceResolverFactory;
+    @Reference
     private DynamicClassLoaderManager dclm;
 
 
@@ -167,12 +110,11 @@ public class JCRHttpCacheStoreImpl extends AbstractJCRCacheMBean<CacheKey, Cache
     }
 
     @Activate
-    protected void activate(ComponentContext context) {
-        Dictionary<?, ?> properties = context.getProperties();
-        cacheRootPath = PropertiesUtil.toString(properties.get(PN_ROOTPATH), DEFAULT_ROOTPATH) + "/" + JCRHttpCacheStoreConstants.ROOT_NODE_NAME;
-        bucketTreeDepth = PropertiesUtil.toInteger(properties.get(PN_BUCKETDEPTH), DEFAULT_BUCKETDEPTH);
-        deltaSaveThreshold = PropertiesUtil.toInteger(properties.get(PN_SAVEDELTA), DEFAULT_SAVEDELTA);
-        expireTimeInSeconds = PropertiesUtil.toInteger(properties.get(PN_EXPIRETIMEINSECONDS), DEFAULT_EXPIRETIMEINSECONDS);
+    protected void activate(Map<String,Object> properties) {
+        cacheRootPath = PropertiesUtil.toString(properties.get(Config.PN_ROOTPATH), Config.DEFAULT_ROOTPATH) + "/" + JCRHttpCacheStoreConstants.ROOT_NODE_NAME;
+        bucketTreeDepth = PropertiesUtil.toInteger(properties.get(Config.PN_BUCKETDEPTH), Config.DEFAULT_BUCKETDEPTH);
+        deltaSaveThreshold = PropertiesUtil.toInteger(properties.get(Config.PN_SAVEDELTA), Config.DEFAULT_SAVEDELTA);
+        expireTimeInMilliSeconds = PropertiesUtil.toLong(properties.get(Config.PN_EXPIRETIMEINMILLISECONDS), Config.DEFAULT_EXPIRETIMEINMILISECONDS);
     }
 
     @Override
@@ -184,9 +126,11 @@ public class JCRHttpCacheStoreImpl extends AbstractJCRCacheMBean<CacheKey, Cache
             final BucketNodeFactory factory = new BucketNodeFactory(session, cacheRootPath, key, bucketTreeDepth);
             final Node bucketNode = factory.getBucketNode();
 
-            final Node entryNode = new BucketNodeHandler(bucketNode, dclm).createOrRetrieveEntryNode(key);
+            final Node entryNode = new BucketNodeHandler(bucketNode, dclm).createOrRetrieveEntryNode(key, expireTimeInMilliSeconds);
 
-            new EntryNodeWriter(session, entryNode, key, content, expireTimeInSeconds).write();
+            long expiryTime = (key.getExpiryForCreation() > 0) ? key.getExpiryForCreation() : expireTimeInMilliSeconds;
+
+            new EntryNodeWriter(session, entryNode, key, content,expiryTime).write();
             session.save();
 
             incrementLoadSuccessCount();
@@ -304,6 +248,11 @@ public class JCRHttpCacheStoreImpl extends AbstractJCRCacheMBean<CacheKey, Cache
     }
 
     @Override
+    public String getStoreType() {
+        return HttpCacheStore.VALUE_JCR_CACHE_STORE_TYPE;
+    }
+
+    @Override
     public void run() {
         purgeExpiredEntries();
     }
@@ -321,7 +270,7 @@ public class JCRHttpCacheStoreImpl extends AbstractJCRCacheMBean<CacheKey, Cache
 
     @Override
     public long getTtl() {
-        return expireTimeInSeconds;
+        return expireTimeInMilliSeconds;
     }
 
     @Override